refactor: move OpeningRegionKeeper to common_meta
