+++ src/common/meta/src/lib.rs
@@ -29,6 +29,7 @@ pub mod kv_backend;
 pub mod metrics;
 pub mod peer;
 pub mod range_stream;
+pub mod region_keeper;
 pub mod rpc;
 pub mod sequence;
 pub mod state_store;
